ch4/ofi: refactor ofi provider code to init_provider.c
The provider selection code is very complex -- ~500 lines. Separate the
code so future work on it can be contained.
